+++ CHANGELOG.rst
@@ -167,3 +167,4 @@ Changelog
 
 * Updated usage documentation for plugins.
 * Added transaction fee option to banking functions.
+* Added tax feature to banking plugin.